requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.31
psycopg[binary]==3.1.19
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
httpx==0.27.0
pydantic==2.7.4
alembic==1.13.1
pytest==8.2.0
